Ignore additional classpath on usecase when using the plugin
It is not appropriate for the plugin to assume that an arbitrary classpath will be present (particularly if used in CI-type pipelines).
In order for the before refactor classes to compile the classpath will already need to include the class path. If the 'after' classpath needs to be different either the project pom can be amended upfront or else the plugin can be declared with additional dependencies.